requirements.txt
streamlit
scikit-learn
nltk
pandas
tmdbv3api
requests
seaborn
beautifulsoup4


flask-pymysql